refactor: Update cloud PC provisioning policy resource type name
Update the `Metadata` method in the `CloudPcProvisioningPolicyResource` struct to set the `TypeName` field to `req.ProviderTypeName + "_graph_cloud_pc_provisioning_policy"`. This change ensures that the resource type name is correctly formatted for the graph.

Refactor the `resource.go` file in the `devicemanagement/v1.0/cloudPcProvisioningPolicy` package to update the resource type name.