refactor: Update client_certificate_file_path attribute description in M365Provider schema
Update the description of the `client_certificate_file_path` attribute in the `M365Provider` schema to clarify that it supports the PKCS#12 (.pfx or .p12) file format. The attribute should contain the certificate, private key, and optionally a certificate chain. If the file is encrypted, the `client_certificate_password` attribute should be used. This change improves the documentation and usability of the attribute.